Update pt-BR.ts
Hello, I am bringing an update on the translation, this time for the letters B and C.

Except for the observations below, I only translated directly what I am sure is correct.

These observations are about changes to items that were translated too literally, and I adapted them to better fit the game.

Beijing: In Brazil, we use the name "Pequim" to refer to this city.
Brickworks: Changed from "Alvenaria" to "Fábrica de tijolos". In Brazil, the former refers to any work with cement or stone walls, while the latter refers to a factory that makes stone blocks or bricks (which, in my opinion, fits the game better).