Added ApplyToAllComponent for Autopilot Settings (#326)
* translations and fixes

* more translations

changed Ironclad from "Panzerung" to "Panzerschiff". Having seen the factory picture, I guess this is what is meant. --> https://de.wikipedia.org/wiki/Panzerschiff

* more translations

* add warning into Atomium

added the warning from
https://discord.com/channels/631551126377857044/1259193266276929598
into Atomium too (currently only in Clone Lab Description)

* add warning into Atomium

added the warning from
https://discord.com/channels/631551126377857044/1259193266276929598
into Atomium too (currently only in Clone Lab Description)

* ISS Description

translated ISS Description

* Added ApplyToAllComponent for Autopilot Settings

Added ApplyToAllComponent for Autopilot Settings
I adapted code from ResourceImportComponent.tsx
This shall solve the following suggestion:
https://discord.com/channels/631551126377857044/1259121503782567997

* more translations

Added ": %{person}" to the SourceGreatPerson as in English version

* more translations

Added ": %{res}" to the PlayerTradeYou*** as in English version

* Clarification on Golden Gate Bridge

Clarifying that the +2 power effect relates to the bridge, not other power plants.

* Fix to St. Basil translation

the "adjacent" was missing in German translation

* Maglev Translation

Asked ingame in DE chat if we prefer rather "Transrapid" which is a brand name or "Magnetschwebebahn" which would be the original translation of "magnetic levitation".
Since we Germans invented this ... we chose Transrapid in chat.

* Clarification on CN Tower

Alter vs. Zeitalter because ingame it is "Zeitalter". -> Change upon request in DE chat (ingame).

* Changed Wonder descriptions

1.) Main change: in DE chat (ingame) improved readability was asked. Thus, like breaks have been added between multiple effects of wonders.

2.) Level and Tier translations were the same ("Stufe"), now Level is "Ausbaustufe" to make the difference clear.

3.) Harmonized "+ xx happiness" in Wonderdescriptions to "+x Zufriedenheit." - There were different versions like "happiness increased by +x" / "+x happiness" / etc.

* More Wonders & Socialism

Socialism was missing the change from "most expensive" to "cheapest" and was not translated at all.

* Fix missing imports

---------

Co-authored-by: Ruoyu Sun <ruoysun@gmail.com>